moved cooldown stuff, added javadoc
